Correct event emission in `_stakeLocked`
Change the event emit value to match the name of the value for the event emitted during `_stakeLocked` function. It's emitting the number of stakes the user has instead of the new locked stake index. Added ` - 1` to the value being emitted to get it to match with the name of the parameter for the event.